# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Dungeons & Databases API
  version: 1.0.0
  description: OpenAPI specification for the Flask backend.
servers:
  - url: /api
...
paths:
...
  /combats:
    post:
      tags: [Combat]
      summary: Start a new dungeon combat for the active character
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Combat created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CombatBootstrapResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /combats/{combatId}:
    parameters:
      - $ref: '#/components/parameters/CombatId'
    get:
      tags: [Combat]
      summary: Fetch one combat state
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Combat state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CombatResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /combats/{combatId}/{action}:
    parameters:
      - $ref: '#/components/parameters/CombatId'
      - $ref: '#/components/parameters/CombatAction'
    get:
      tags: [Combat]
      summary: Resolve a combat action for the requested combat
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Combat resolved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CombatOutcomeResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    CombatId:
      name: combatId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
    CombatAction:
      name: action
      in: path
      required: true
      schema:
        type: string
        enum: [attack, next_combat, run, go_home]
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          oneOf:
            - type: string
            - type: array
              items: {}
      required: [error]
...
    CharacterResponse:
      type: object
      required:
        - id
        - user_id
        - name
        - level
        - experience
        - experience_to_next_level
        - max_health
        - health
        - damage
        - state
        - bonus_health
        - bonus_damage
      properties:
        id:
          type: integer
        user_id:
          type: integer
        name:
          type: string
        level:
          type: integer
        experience:
          type: integer
        experience_to_next_level:
          type: integer
        max_health:
          type: integer
        health:
          type: integer
        damage:
          type: integer
        state:
          type: string
          enum: [HOME, DUNGEON_COMBAT, DUNGEON_VICTORY]
        bonus_health:
          type: integer
        bonus_damage:
          type: integer
...
    CombatResponse:
      type: object
      required:
        - id
        - character_id
        - character_health
        - enemy
      properties:
        id:
          type: integer
        character_id:
          type: integer
        character_health:
          type: integer
        enemy:
          $ref: '#/components/schemas/CombatEnemyResponse'
    CombatEnemyResponse:
      type: object
      required:
        - type_id
        - name
        - description
        - level
        - health
        - max_health
        - damage
        - base_health
        - base_damage
      properties:
        type_id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        level:
          type: integer
        health:
          type: integer
        max_health:
          type: integer
        damage:
          type: integer
        base_health:
          type: integer
        base_damage:
          type: integer
...
    CombatBootstrapResponse:
      type: object
      required: [combat, character]
      properties:
        combat:
          $ref: '#/components/schemas/CombatResponse'
        character:
          $ref: '#/components/schemas/CharacterResponse'
    DroppedItem:
      type: object
      required: [id, name, slot_type, health, damage, level]
      properties:
        id:
          type: string
        name:
          type: string
        slot_type:
          type: string
          enum: [weapon, shield, armor, helmet, ring, necklace]
        health:
          type: integer
        damage:
          type: integer
        level:
          type: integer
    CombatOutcomeResponse:
      type: object
      required:
        - character
        - combat
        - message
        - victory
        - items_dropped
        - player_died
        - success
      properties:
        character:
          $ref: '#/components/schemas/CharacterResponse'
        combat:
          $ref: '#/components/schemas/CombatResponse'
          nullable: true
        message:
          type: string
        victory:
          type: boolean
        items_dropped:
          type: array
          items:
            $ref: '#/components/schemas/DroppedItem'
        player_died:
          type: boolean
        success:
          type: boolean
        dice_roll:
          type: integer
          nullable: true
        damage:
          type: integer
          nullable: true